requirements.txt
streamlit
requests
gtts
pillow
openai
python-dotenv